requirements.txt
cheroot==8.5.2
jaraco.functools==3.2.1
more-itertools==8.7.0
mysql-connector-python==8.0.23
pkg-resources==0.0.0
protobuf==3.15.3
python-decouple==3.4
six==1.15.0
web.py==0.61
